+++ lib/src/pod_list_builder.dart
@@ -20,13 +20,13 @@ import '/xyz_pod.dart';
 ///
 /// This widget is useful for managing and responding to the state changes of multiple
 /// `Pod` objects within the Flutter UI.
-class PodListBuilder extends StatefulWidget {
+class PodListBuilder<T extends dynamic> extends StatefulWidget {
   //
   //
   //
 
   /// A list of `Pod` objects. The widget listens to changes in these objects.
-  final List<Pod?> pods;
+  final List<Pod<T>?> pods;
 
   //
   //
@@ -42,7 +42,7 @@ class PodListBuilder extends StatefulWidget {
   /// A builder function that takes the current `BuildContext`, the child `Widget`
   /// and a list of current values of the `pods`. It returns a `Widget` that is
   /// rebuilt every time one of the `Pod` objects notifies its listeners.
-  final Widget Function(BuildContext, Widget?, List) builder;
+  final Widget Function(BuildContext, Widget?, List<T?>) builder;
 
   //
   //
@@ -57,10 +57,31 @@ class PodListBuilder extends StatefulWidget {
   /// - `child` (optional): A child widget to be passed to the builder function.
   const PodListBuilder({
     super.key,
-    this.pods = const [],
+    this.pods = const <Pod<T>?>[],
     required this.builder,
     this.child,
   });
+
+  //
+  //
+  //
+
+  /// Constructs a `PodListBuilder` widget.
+  ///
+  /// Parameters:
+  /// - `key`: An identifier for this widget in the widget tree.
+  /// - `pods`: A list of `Pod` objects to listen to.
+  /// - `builder`: A function that builds the UI based on the `pods`.
+  factory PodListBuilder.values({
+    Key? key,
+    List<Pod<T>> pods = const [],
+    required Widget Function(List<T?>) builder,
+  }) {
+    return PodListBuilder<T>(
+      pods: pods,
+      builder: (_, __, values) => builder(values),
+    );
+  }
 
   //
   //
@@ -72,7 +93,7 @@ class PodListBuilder extends StatefulWidget {
 
 // ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
 
-class PodListBuilderState extends State<PodListBuilder> {
+class PodListBuilderState<T extends dynamic> extends State<PodListBuilder<T>> {
   //
   //
   //
@@ -91,7 +112,7 @@ class PodListBuilderState extends State<PodListBuilder> {
   //
 
   @override
-  void didUpdateWidget(PodListBuilder oldWidget) {
+  void didUpdateWidget(PodListBuilder<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
     // Remove listeners from old Pods and add listeners to new Pods.
     for (final oldPod in oldWidget.pods) {
@@ -120,7 +141,11 @@ class PodListBuilderState extends State<PodListBuilder> {
   @override
   Widget build(BuildContext context) {
     // Build the widget using the provided builder function.
-    return widget.builder(context, widget.child, widget.pods.map((e) => e?.value).toList());
+    return widget.builder(
+      context,
+      widget.child,
+      widget.pods.map((e) => e?.value).toList(),
+    );
   }
 
   //